remove grub2-efi transformation
All tools and support files are now the same for bios and efi - only the actual
bootloader is different. There is thus no reason to build any tools for efi,
but there is no way to avoid it. We just build and install them and overwrite
with bios files.

By using the same transformation we avoid the problems with the transformation
being included in some un-transformed files.

The old grub2-efi specific are kept as symlinks for backward compatibility.
grub2-efi is still very much work-in-progress and the grub2-efi could probably
be dropped very soon. It would also make sense to move all shared files to a
sub-package.