Revert "Add patch from lersek to fix UEFI Windows dual-boot - #1347291"
This reverts commit 44da1e072bc0889caf0cfdf01d9e0b9b0a1036a6.
Reverting for now as there's some unresolved questions that
make me unwilling to go through with this change - like why
ppc64 build fails on Rawhide, and whether the changes from
the f24 branch that aren't on master or f25 should be brought
forward. So reverting to the status quo until I can talk to
Peter.